Read data in stored in a file
Goal of the Code:
The goal of this code is to read integer data stored in a file named "N_numbers.txt" and display each number on the console. It demonstrates basic file handling operations in C++:

Opening a File: The inFile.open("N_numbers.txt"); statement opens "N_numbers.txt" file in the current directory for reading.
Reading from a File: The while (inFile >> number) loop reads integers from the file until the end-of-file (EOF) is reached.
Outputting Data: Each integer read is printed with a message "This is the data read :" followed by the number.