read, write file
The goal of this code is to manage and manipulate data for a list of football players. The program allows the user to read player data from a file, display player information, update player data, and write updated data back to a file. The key functionalities provided by the program include reading input data, displaying information, updating player details, and writing the data back to an output file.