LinkedList class
The goal of this code is to demonstrate the functionality of a LinkedList class by allowing the user to interact with a linked list through various operations such as appending nodes, inserting nodes, deleting nodes, printing the list, and printing the list in reverse order. The user can perform these actions via a simple text-based menu.